backend/main.py
"""
Babywise Chatbot - Main FastAPI Application

This module initializes the FastAPI application and defines the API endpoints
for the Babywise Chatbot.
"""

import os
import logging
from typing import Dict, Optional
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
)
logger = logging.getLogger(__name__)

# Initialize FastAPI app
app = FastAPI(
    title="Babywise Chatbot API",
    description="API for the Babywise Chatbot, a domain-specific assistant for baby care guidance",
    version="1.0.0",
)

# Add CORS middleware to allow cross-origin requests from the frontend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # In production, replace with specific origins
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Define request and response models
class ChatRequest(BaseModel):
    """Model for chat request data."""
    message: str
    thread_id: str
    language: str = "en"

class ChatResponse(BaseModel):
    """Model for chat response data."""
    response: str
    context: Dict
    metadata: Dict

# Health check endpoint
@app.get("/health")
async def health_check():
    """Health check endpoint to verify the API is running."""
    return {"status": "healthy"}

# Chat endpoint
@app.post("/chat", response_model=ChatResponse)
async def chat(request: ChatRequest):
    """
    Process a chat message and return a response.
    
    This endpoint will be connected to the LangChain/LangGraph workflow
    in future implementations.
    """
    try:
        # Placeholder for actual chat processing logic
        # This will be replaced with the LangChain/LangGraph workflow
        logger.info(f"Received message: {request.message} (Thread: {request.thread_id}, Language: {request.language})")
        
        # For now, return a simple response
        return ChatResponse(
            response="This is a placeholder response. The Babywise Chatbot is under development.",
            context={"detected_domain": "general"},
            metadata={"timestamp": "2023-01-01T00:00:00Z"}
        )
    except Exception as e:
        logger.error(f"Error processing chat request: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

# Events endpoint (for routine tracking)
@app.post("/events")
async def create_event():
    """
    Create a new routine event (e.g., sleep, feeding).
    
    This endpoint will be implemented in future phases.
    """
    return {"status": "not implemented"}

@app.get("/events")
async def get_events():
    """
    Retrieve routine events.
    
    This endpoint will be implemented in future phases.
    """
    return {"status": "not implemented"}

# Summary endpoint
@app.get("/summary")
async def get_summary():
    """
    Generate a summary report of routine events.
    
    This endpoint will be implemented in future phases.
    """
    return {"status": "not implemented"}

if __name__ == "__main__":
    # This block is used when running the app directly with Python
    # For production, use: uvicorn main:app
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000) 