backend/requirements.txt
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.0.5
langgraph>=0.2.20,<0.3
langsmith>=0.1.0
openai>=1.6.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.6
typing-extensions>=4.9.0
aiohttp>=3.9.0
async-timeout>=4.0.0
sqlalchemy==2.0.27
aiosqlite==0.19.0
starlette==0.36.3
aiohttp==3.9.3
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pymongo[srv]>=4.6.1
motor>=3.3.2
dnspython>=2.5.0
certifi>=2024.2.2